[Decomp][GameObject] Cohesion split: extract destructible building to GameObjectDestructible.cpp
